Move GeoTrellis orphan instances to new `bridge` package
- Previously they were defined in `datamodel`, even though any section
  of RF code could need them in theory. In implementing the Masking
  AST operation, we need to import these instances in `tool`. To avoid
  circular deps, we have to move them out into a new package.
