+++ app-backend/bridge/build.sbt
@@ -1,0 +1,5 @@
+name := "raster-foundry-bridge"
+
+scalacOptions ++= Seq("-feature", "-unchecked", "-deprecation")
+
+addCompilerPlugin("org.scalamacros" % "paradise" % "2.1.0" cross CrossVersion.full)
