+++ app-backend/datamodel/src/main/scala/com/azavea/rf/datamodel/Scene.scala
@@ -3,6 +3,7 @@ package com.azavea.rf.datamodel
 import java.sql.Timestamp
 import java.util.UUID
 
+import com.azavea.rf.bridge._
 import geotrellis.vector.Geometry
 import geotrellis.slick.Projected
