+++ app-backend/tool/src/main/scala/ast/codec/MapAlgebraOperationCodecs.scala
@@ -2,6 +2,7 @@ package com.azavea.rf.tool.ast.codec
 
 import com.azavea.rf.tool.ast._
 import com.azavea.rf.tool.eval._
+import com.azavea.rf.bridge._
 import io.circe._
 import io.circe.syntax._
